refactor: move error types to appropiate submodules
