+++ crates/rust-client/src/notes/note_screener.rs
@@ -1,12 +1,17 @@
 use alloc::{collections::BTreeSet, rc::Rc};
 use core::fmt;
 
-use miden_objects::{accounts::AccountId, assets::Asset, notes::Note, Word};
+use miden_objects::{
+    accounts::AccountId,
+    assets::Asset,
+    notes::{Note, NoteId},
+    AccountError, AssetError, Word,
+};
 use winter_maybe_async::{maybe_async, maybe_await};
 
 use crate::{
-    errors::{InvalidNoteInputsError, NoteScreenerError},
-    store::Store,
+    notes::{InvalidNoteInputsError, NoteScreenerError},
+    store::{Store, StoreError},
     transactions::known_script_roots::{P2ID, P2IDR, SWAP},
 };
 
@@ -189,5 +194,75 @@ impl<S: Store> NoteScreener<S> {
             .iter()
             .map(|account_id| (*account_id, NoteRelevance::Always))
             .collect())
+    }
+}
+
+// NOTE SCREENER ERROR
+// ================================================================================================
+
+/// Error when screening notes to check relevance to a client
+#[derive(Debug)]
+pub enum NoteScreenerError {
+    InvalidNoteInputsError(InvalidNoteInputsError),
+    StoreError(StoreError),
+}
+
+impl From<InvalidNoteInputsError> for NoteScreenerError {
+    fn from(error: InvalidNoteInputsError) -> Self {
+        Self::InvalidNoteInputsError(error)
+    }
+}
+
+impl From<StoreError> for NoteScreenerError {
+    fn from(error: StoreError) -> Self {
+        Self::StoreError(error)
+    }
+}
+
+impl fmt::Display for NoteScreenerError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            NoteScreenerError::InvalidNoteInputsError(note_inputs_err) => {
+                write!(f, "error while processing note inputs: {note_inputs_err}")
+            },
+            NoteScreenerError::StoreError(store_error) => {
+                write!(f, "error while fetching data from the store: {store_error}")
+            },
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum InvalidNoteInputsError {
+    AccountError(NoteId, AccountError),
+    AssetError(NoteId, AssetError),
+    NumInputsError(NoteId, usize),
+    BlockNumberError(NoteId, u64),
+}
+
+impl fmt::Display for InvalidNoteInputsError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            InvalidNoteInputsError::AccountError(note_id, account_error) => {
+                write!(f, "account error for note with ID {}: {account_error}", note_id.to_hex())
+            },
+            InvalidNoteInputsError::AssetError(note_id, asset_error) => {
+                write!(f, "asset error for note with ID {}: {asset_error}", note_id.to_hex())
+            },
+            InvalidNoteInputsError::NumInputsError(note_id, expected_num_inputs) => {
+                write!(
+                    f,
+                    "expected {expected_num_inputs} note inputs for note with ID {}",
+                    note_id.to_hex()
+                )
+            },
+            InvalidNoteInputsError::BlockNumberError(note_id, read_height) => {
+                write!(
+                    f,
+                    "note input representing block with value {read_height} for note with ID {}",
+                    note_id.to_hex()
+                )
+            },
+        }
     }
 }
