+++ crates/rust-client/src/store/note_record/output_note_record.rs
@@ -8,7 +8,7 @@ use miden_objects::{
 };
 
 use super::{InputNoteRecord, NoteRecordDetails, NoteStatus};
-use crate::errors::ClientError;
+use crate::ClientError;
 
 // OUTPUT NOTE RECORD
 // ================================================================================================
