+++ crates/rust-client/src/store/sqlite_store/migrations.rs
@@ -2,7 +2,7 @@ use lazy_static::lazy_static;
 use rusqlite::Connection;
 use rusqlite_migration::{Migrations, M};
 
-use crate::errors::StoreError;
+use crate::store::StoreError;
 
 // MIGRATIONS
 // ================================================================================================
